requirements.txt
pypdf2
pyyaml